refactor: use routes.php as single source of truth; register expo macro via Route::macro instead of mixin
refs: https://github.com/RonasIT/laravel-exponent-push-notifications/issues/28
