+++ src/Http/routes.php
@@ -1,15 +1,12 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use NotificationChannels\ExpoPushNotifications\ExpoRouter;
 use NotificationChannels\ExpoPushNotifications\Http\ExpoController;
 
-Route::prefix(ExpoRouter::DEFAULT_PREFIX)
-    ->middleware(ExpoRouter::DEFAULT_MIDDLEWARE)
+Route::prefix('exponent/devices')
+    ->middleware('expo.middleware')
+    ->controller(ExpoController::class)
     ->group(function () {
-        Route::controller(ExpoController::class)->group(function () {
-            foreach (ExpoRouter::ROUTES as $route) {
-                Route::{$route['method']}($route['uri'], $route['action']);
-            }
-        });
+        Route::post('subscribe', 'subscribe');
+        Route::post('unsubscribe', 'unsubscribe');
     });
